requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
langgraph
langchain-groq
langchain-core
python-dotenv
psycopg2-binary
